+++ lib/brightpearl/resources.rb
@@ -4,6 +4,9 @@ require 'brightpearl/resource'
 require 'brightpearl/resources/contact'
 require 'brightpearl/resources/contact_custom_field'
 require 'brightpearl/resources/contact_custom_field_metadata'
+require 'brightpearl/resources/contact_group'
+require 'brightpearl/resources/contact_tag'
+require 'brightpearl/resources/lead_source'
 
 # Orders
 require 'brightpearl/resources/order'
@@ -13,6 +16,8 @@ require 'brightpearl/resources/order_status'
 require 'brightpearl/resources/order_status_update'
 require 'brightpearl/resources/order_custom_field'
 require 'brightpearl/resources/order_custom_field_metadata'
+require 'brightpearl/resources/order_stock_status'
+require 'brightpearl/resources/order_type'
 
 require 'brightpearl/resources/goods_out_note'
 
@@ -29,6 +34,7 @@ require 'brightpearl/resources/channel'
 require 'brightpearl/resources/product_category'
 require 'brightpearl/resources/product_type'
 require 'brightpearl/resources/warehouse'
+require 'brightpearl/resources/shipping_method'
 
 # Integration
 require 'brightpearl/resources/webhook'
@@ -36,3 +42,5 @@ require 'brightpearl/resources/webhook'
 # Accounting
 require 'brightpearl/resources/tax_code'
 require 'brightpearl/resources/customer_payment'
+require 'brightpearl/resources/nominal_code'
+require 'brightpearl/resources/payment_method'