+++ lib/brightpearl/resources/order_stock_status.rb
@@ -1,0 +1,21 @@
+module Brightpearl
+  class OrderStockStatus < Resource
+    extend Brightpearl::APIOperations::Get
+    class << self
+      def resource_path
+        "order-service/order-stock-status"
+      end
+
+      def get(id_set = nil)
+        if id_set
+          super
+        else
+          super(nil)
+        end
+      end
+
+      # https://api-docs.brightpearl.com/order/order-stock-status/get.html
+
+    end
+  end
+end